requirements.txt
absl-py==2.1.0
accelerate==0.26.1
addict==2.4.0
aiofiles==23.2.1
altair==5.2.0
annotated-types==0.6.0
anyio==4.2.0
appdirs==1.4.4
arcface==0.0.8
argon2-cffi-bindings==21.2.0
argon2-cffi==23.1.0
arrow==1.3.0
astropy-iers-data==0.2024.3.18.0.29.47
astropy==6.0.0
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
attrs==23.2.0
babel==2.14.0
basicsr==1.4.2
beautifulsoup4==4.12.3
bleach==6.1.0
blis==0.7.11
bottleneck==1.3.7
brisque==0.0.16
brotli==1.1.0
cake==0.2.3
catalogue==2.0.10
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
clean-fid==0.1.35
click==8.1.7
cloudpathlib==0.16.0
cmake==3.29.2
colorama==0.4.6
comm==0.2.1
confection==0.1.4
contourpy==1.2.0
customtkinter==5.2.2
cycler==0.12.1
cymem==2.0.8
darkdetect==0.8.0
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
diffusers==0.25.1
docker-pycreds==0.4.0
einops==0.7.0
exceptiongroup==1.2.0
executing==2.0.1
facexlib==0.3.0
fastai==1.0.60
fastapi==0.110.0
fastjsonschema==2.19.1
fastprogress==1.0.3
ffmpeg-python==0.2.0
ffmpeg==1.4
ffmpy==0.3.2
filelock==3.13.1
filterpy==1.4.5
flatbuffers==24.3.7
fonttools==4.47.2
fqdn==1.5.1
fsspec==2023.12.2
ftfy==6.2.0
future==0.18.3
gast==0.5.4
gdown==5.0.0
gfpgan==1.3.8
gitdb==4.0.11
gitpython==3.1.41
google-pasta==0.2.0
gradio-client==0.12.0
gradio==4.21.0
grpcio==1.60.0
h11==0.14.0
h5py==3.10.0
httpcore==1.0.4
httpx==0.27.0
huggingface-hub==0.20.3
idna==3.6
image-quality==1.2.7
imageio==2.33.1
imgaug==0.4.0
importlib-metadata==7.0.1
importlib-resources==6.1.1
ipykernel==6.29.0
ipython==8.18.1
ipywidgets==8.1.1
isoduration==20.11.0
jedi==0.19.1
jinja2==3.1.3
joblib==1.4.2
json5==0.9.14
jsonpointer==2.4
jsonschema-specifications==2023.12.1
jsonschema==4.21.1
jupyter-client==8.6.0
jupyter-core==5.7.1
jupyter-events==0.9.0
jupyter-lsp==2.2.2
jupyter-server-terminals==0.5.2
jupyter-server==2.12.5
jupyterlab-pygments==0.3.0
jupyterlab-server==2.25.2
jupyterlab-widgets==3.0.9
jupyterlab==4.0.11
keras==3.1.1
kiwisolver==1.4.5
langcodes==3.3.0
lazy-loader==0.3
libclang==18.1.1
libsvm-official==3.32.0
libsvm==3.23.0.4
llvmlite==0.41.1
lmdb==1.4.1
lpips==0.1.4
markdown-it-py==3.0.0
markdown==3.5.2
markupsafe==2.1.4
matplotlib-inline==0.1.6
matplotlib==3.8.2
mdurl==0.1.2
mistune==3.0.2
ml-dtypes==0.3.2
mpmath==1.3.0
murmurhash==1.0.10
mutagen==1.47.0
namex==0.0.7
nbclient==0.9.0
nbconvert==7.14.2
nbformat==5.9.2
nest-asyncio==1.6.0
networkx==3.2.1
notebook-shim==0.2.3
numba==0.58.1
numexpr==2.8.8
numpy==1.24.4
nvidia-ml-py3==7.352.0
openai-clip==1.0.1
opencv-python==4.9.0.80
opt-einsum==3.3.0
optree==0.10.0
orjson==3.9.15
outcome==1.3.0.post0
overrides==7.6.0
packaging==23.2
pandas==2.2.0
pandocfilters==1.5.1
parso==0.8.3
patch==1.16
pillow==9.3.0
pip==24.0
platformdirs==4.1.0
preshed==3.0.9
prometheus-client==0.19.0
prompt-toolkit==3.0.43
protobuf==4.25.2
psutil==5.9.8
pure-eval==0.2.2
pycparser==2.21
pycryptodomex==3.20.0
pydantic-core==2.14.6
pydantic==2.5.3
pydub==0.25.1
pyerfa==2.0.1.1
pygments==2.17.2
pyiqa==0.1.10
pyparsing==3.1.1
pysocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.1
python-io==0.3
python-json-logger==2.0.7
python-multipart==0.0.9
pytorch-fid==0.3.0
pytz==2023.3.post1
pywin32==306
pywinpty==2.0.12
pyyaml==6.0.1
pyzmq==25.1.2
realesrgan==0.3.0
referencing==0.32.1
regex==2023.12.25
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rpds-py==0.17.1
ruff==0.3.2
safetensors==0.4.2
scikit-image==0.22.0
scikit-learn==1.5.0
scipy==1.9.1
selenium==4.18.1
semantic-version==2.10.0
send2trash==1.8.2
sentry-sdk==1.39.2
setproctitle==1.3.3
setuptools==56.0.0
shapely==2.0.3
shellingham==1.5.4
six==1.16.0
smart-open==6.4.0
smmap==5.0.1
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
spacy==3.7.2
srsly==2.4.8
stack-data==0.6.3
starlette==0.36.3
sympy==1.12
tb-nightly==2.16.0a20240123
tensorboard-data-server==0.7.2
tensorboard==2.16.2
tensorboardx==2.6.2.2
tensorflow-hub==0.16.1
tensorflow-intel==2.16.1
tensorflow-io-gcs-filesystem==0.31.0
tensorflow==2.16.1
termcolor==2.4.0
terminado==0.18.0
tf-keras-nightly==2.16.0.dev2024012310
tf-keras==2.16.0
thinc==8.2.2
threadpoolctl==3.5.0
tifffile==2023.12.9
timm==0.9.16
tinycss2==1.2.1
tokenizers==0.15.1
tomli==2.0.1
tomlkit==0.12.0
toolz==0.12.1
torch==2.2.1+cu121
torchaudio==2.2.1+cu121
torchvision==0.17.1
tornado==6.4
tqdm==4.66.1
traitlets==5.14.1
transformers==4.37.1
trio-websocket==0.11.1
trio==0.24.0
typer==0.9.0
types-python-dateutil==2.8.19.20240106
typing-extensions==4.9.0
tzdata==2023.4
uri-template==1.3.0
urllib3==2.1.0
uvicorn==0.27.1
wandb==0.16.2
wasabi==1.1.2
wcwidth==0.2.13
weasel==0.3.4
webcolors==1.13
webdriver-manager==4.0.1
webencodings==0.5.1
websocket-client==1.7.0
websockets==11.0.3
werkzeug==3.0.1
wheel==0.43.0
widgetsnbextension==4.0.9
wrapt==1.16.0
wsproto==1.2.0
yapf==0.40.2
yt-dlp==2023.12.30
zipp==3.17.0
